Add support for service reconfigure
In support of PaaS-style use-cases where CFME services
are utilized to drive automation-driven provisioning, being
able to reconfigure an existing service with new parameters
is a requirement.  In support of this, services can now be
flagged as reconfigurable, which will give the user the
ability to modify the service parameters via the service
dialog.  Once submitted, the reconfiguration request will
use the existing approval process before invoking the assigned
custom automation to perform the reconfiguration steps.

https://bugzilla.redhat.com/show_bug.cgi?id=1059834