Updated the valid states for provision requests
The provision task has a state named "provisioned" which is
not currently a valid state for its associated provision request.
As a result, after a provision task has finished provisioning a
VM, the task's status can no longer be bubbled up to its parent
provision request due to a validation error.

https://bugzilla.redhat.com/show_bug.cgi?id=1122123