Remove temporary code for comparing C-binding with FFI-binding
VixDiskLib.rb in this current incarnation simply tested an environment
variable and then loaded the new FFI-binding to the library (the
default) or loaded the old C-language binding to VixDisklib.  We will
be removing access to the C-language binding so this file is no longer
needed.  Note that the next commit will rename VixDiskLibClient.rb,
which VixDiskLib.rb required in order to start the FFI-binding, back
to VixDiskLib.rb.