Replace temporary code FFI-binding code
The previous commit removed the file VixDiskLib.rb, which loaded either
the VixDiskLib FFI-binding or the C-language binding, based on an
environment variable.  We are removing access to the C-binding, so the
file VixDiskLibClient.rb, which starts the FFI-binding, is being renamed
to VixDiskLib.rb, so that the FFI-binding will always be used.