Use tenant in openstack provisioning
Update MiqProvisionOpenstack::Cloning to use the cloud_tenant selected during
the provisioning process.  The cloud_tenant should be used in establishing the
openstack connection in order to specify which tenant the newly created instance
should reside.

The caveat here is that the EMS user must be associated to the selected tenant
in Openstack.  If the EMS user is not associated to the selected tenant, the
provisioning will fail with an unauthorized error from Openstack.