Change AMQP default port to non-ssl
We have gone back and forth on this a bit.  But, I believe it's shaken itself
out during testing.

When setting up AMQP to use SSL, several steps are involved including copying
ssl certs from the AMQP server over to the ManageIQ appliance.  Defaulting to
the SSL AMQP port assumes that the user is going to take these steps.  It's
better to assume non-ssl, and then include changing the default AMQP port number
part of the steps for configuring AMQP to use SSL in the appliance.

https://bugzilla.redhat.com/show_bug.cgi?id=1127954