Remove a sporadic failing spec.
It's useful to verify that a schedule's last_run_on is updated only on completion and not on queueing of the scheduled work.

Unfortunately, this test relies on the underlying implementation of rufus.

If this test is useful to have, we'll have to reorganize the code so we can test just queueing in isolation.