Strip off any leading/trailing whitespace such as a newline.
https://bugzilla.redhat.com/show_bug.cgi?id=1126071

We shouldn't require the VERSION file to be absent of all leading/trailing whitespace such as a trailing
newline at the end of the file as many editors will insert this and it's easy to do this without realizing it matters.

Adding a newline causes the comparison of the rpm version information to NOT match the
VERSION provided by the contents of the smartproxy executable.

This causes the product_updates table to NOT be seeded with the smartproxy found in the filesystem.
:bomb: