Change MiqApproval#approver to use a nil user when roles should be used.
Previously, approver could be set to a UiTaskSet, which meant that
anyone with the approver role could be an approver.  This is now handled
by setting approver to nil, making approver just represent a User
instead of both a User and a UiTaskSet.

Since UiTaskSet is no longer used technically, we instead verify only
against the miq_request_approval MiqUserRole, which
MiqApproval#authorized? had already handled.

Fixes #473
https://bugzilla.redhat.com/show_bug.cgi?id=1130146