Handle Openstack 404 Errors
Openstack can be configured to return 404 errors for certain API requests.  The
one that has been encountered is when Openstack Neutron is configured with the
NoOpFirewall handler.  In this case, calls to get_security_groups will return a
404 HTTP error.

ManageIQ codebase does not handle non-successful API HTTP return codes.  This is
an attempt to simply ignore (and log) any 404 return codes.

https://bugzilla.redhat.com/show_bug.cgi?id=1140191