SCVMM - We should only attempt to gather IP addresses from running VMs.
The IP addresses of the VMs in SCVMM are exposed by the Hyperv server. To enable this:
1) The VM must be running
2) The Data Exchange Service must be running on the VM.

If either of the above are false an error is thrown by Powershell and logged in the evm.log file making for a messy and unreadable file.
This PR will significantly reduce the volume of these errors by checking if the VM is in a running state before we try to get the IP addresses.
We will still print out the same error for VMs that do not have 2) above enabled but only for a limited number of VMs since Integration Services are installed by default.

Note -
We can turn off this error completely but i think customers will find the logging helpful.