Set raw power state in pxe/iso provision
The power_state attribute on the VM was recently changed to read-only.  However,
pxe and iso provisioning in RHEV and VMware still set the power_state after
launch.

This is now changed to set the raw_power_state.  However, there's a bit of a
smell here because we're setting the "raw_power_state" to a non-raw value.

Can't think of a better way right now.

https://bugzilla.redhat.com/show_bug.cgi?id=1149195