Fixed issue - buttons do not light up after a selection is made
This issue was seen in the Provisioning dialog page where despite a
template selection, the Continue button would not light up.

Recent prototype -> jQuery changes exposed this issue (which is an issue
for jQuery, but not for prototype) where if there are multiple elements
in a html page with the same id , for e.g. -
<span id='buttons_on'>
<div id='buttons_on'>
then a jQuery call - $j('#buttons_on') would pick only the first element
, thus introducing a bug where the second element with the same id would
not be handled at all.
On the Provisioning dialog page, the quick_search bar had its own set
of buttons that were causing the above issue with the Provisioning
dialog buttons. The quick_search bar buttons have now been renamed to
quick_search_buttons_on/off to avoid the id clash.