For MiqAeField and MiqAeValue use build+save instead of create!
This PR is based on the Active Record/Postgres.

For Git based Automate model the MiqAeField and MiqAeValue are saved
in the class and instance yaml files and are not separate tables like
in ActiveRecord. Build followed by a save allows us to write just once
into the file as opposed to writing the file several times once a value
or field is saved.