Language selection in webui
This change allows for selection of webui language on a user
and a server level. Should the user settings be set to default,
server settings will apply. Should the server settings be set
to default, webui will set the language based on browser
preferences.

This change also introduces a new translation string 'locale_name',
which is meant to be translated into a locale name in every
supported native language.

Fixes #1008