Fix provision status issue.
Provision status method returns incorrect value.
Will be deprecating the status method in favor of new
statemachine_task_status. Added DEPRECATION warning log message and
fixed issue in miq_provision status method where the result for a
finished or provisioned state is determined solely on the existence
of the vm.

https://bugzilla.redhat.com/show_bug.cgi?id=1025761