eager load association
Associations are not eagerly loaded in Rails 4.  This code assumed that
`disconnects` would be eagerly loaded.  In Rails 4 it was not eagerly
loaded, and would mutate records that had been added to the association
but were not supposed to be mutate.

This commit loads the association and keeps the list in memory so that
only the correct records will be mutated