kubernetes: add event catcher implementation
This patch introduces the initial implementation of the kubernetes event
catcher.

For simplicity it watches the single event entity (instead of the
separates entities: nodes, pods, replication-controller, services) to
trigger a full inventory pull.

In the future, in order to scale, we should watch the specific entities
and receive the relevant information that can be used to update the
inventory (avoiding the repeated full inventory pulls).

Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>